ConversationalBot/requirements.txt
pymongo==4.10.1
python-telegram-bot==21.6